Increase watch deployment timeout
From 20 minutes to 40.

This could maybe be totally removed, added as a workflow defined
param and removed and the timeout added in each workflow as a timeout on
the step, but for now this is a decent safety net.

Just increase it so that it should never get tripped, but has enough
slack that if something minor goes wrong, we won't timeout.

Enough time for a couple of full deploys if queuing.